feat: EditStudyModalButton 생성하여 공용 컴포넌트인 PasswordModalButton 쓸수 있도록 구현
