--- frontend/src/components/edit-study-modal/EditStudyModalController.jsx
@@ -1,15 +1,0 @@
-import { useState } from 'react';
-
-export const useEditStudyModalController = () => {
-  const [editStudyModal, setEditStudyModal] = useState(false);
-
-  const handleOpenEditStudyModal = () => {
-    setEditStudyModal(true);
-  };
-
-  const handleCloseEditStudyModal = () => {
-    setEditStudyModal(false);
-  };
-
-  return { editStudyModal, handleOpenEditStudyModal, handleCloseEditStudyModal };
-};
