+++ frontend/src/pages/study-detail/StudyDetailNav.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import styles from "./StudyDetailNav.module.css";
-import EditStudyModalButton from "@components/edit-study-modal/EditStudyModalButton";
 import StudyEmoji from "./study-emoji/StudyEmoji";
 import DeleteStudyModalButton from "@components/delete-study-modal/DeleteStudyModalButton";
 import ShareModal from "./components/ShareModal";
+import EditStudyModalButton from "./components/EditStudyModalButton";
 
 function StudyDetailNav() {
   const [shareModalOpen, setShareModalOpen] = useState(false);
@@ -23,10 +23,7 @@ function StudyDetailNav() {
           공유하기
         </button>
         <span className={styles.divider}>|</span>
-        <EditStudyModalButton
-          buttonText="수정하기"
-          className={styles.editLink}
-        />
+        <EditStudyModalButton buttonText="수정하기" />
         <span className={styles.divider}>|</span>
         <DeleteStudyModalButton buttonText="스터디 삭제하기" />
