+++ frontend/src/pages/study-detail/components/EditStudyModalButton.module.css
@@ -1,0 +1,13 @@
+.editStudyModalButton {
+  color: var(--color-green-300);
+  cursor: pointer;
+  text-decoration: none;
+  font-weight: 500;
+  font-size: 12px;
+}
+
+@media screen and (min-width: 744px) {
+  .editStudyModalButton {
+    font-size: 16px;
+  }
+}
